Split data/api.js monolith into per-entity modules
Replaces the 425-line src/data/api.js with a folder per concern under
src/data/api/, then re-exports everything from index.js so existing
'from "../../data/api"' imports keep working without touching any
component:

  data/api/
    index.js        barrel re-exports the public surface
    constants.js    ROLES, USER_STATUSES, CHALLENGE_STATUSES, ACTIONS
    helpers.js      unwrap, resetDemoData
    users.js        users + invite serverless call
    challenges.js   challenges + challenge_actions + challenge_participants
    actions.js      actions catalog + by-challenge query
    participation.js participation reads + distinct-participant counts
    groups.js       departments table aliased as Groups
    presets.js      presets + preset_actions
    templates.js    templates table (no UI yet, see #7)
    activityLogs.js activity_logs reads + logActivity
    leaderboard.js  cross-table point aggregations

Every module gets a header docblock and JSDoc on its exported
functions per the Cohesive Comment Standard. Refs #42.
