+++ src/admin-app/src/data/api/constants.js
@@ -1,0 +1,41 @@
+/**
+ * Domain enums shared across the app.
+ *
+ * Strings here mirror the values stored in Postgres. Treat them as
+ * append-only -- adding new values is safe, renaming existing ones
+ * requires a migration.
+ */
+
+export const ROLES = {
+  SUPER_ADMIN: "SuperAdmin",
+  ADMIN: "Admin",
+  GENERAL_USER: "GeneralUser",
+};
+
+export const USER_STATUSES = {
+  ACTIVE: "Active",
+  DEACTIVATED: "Deactivated",
+};
+
+export const CHALLENGE_STATUSES = {
+  ACTIVE: "Active",
+  UPCOMING: "Upcoming",
+  COMPLETED: "Completed",
+  ARCHIVED: "Archived",
+};
+
+/**
+ * Action category taxonomy adopted from MPCA's GreenStep program.
+ * The first item is the catch-all default.
+ */
+export const ACTIONS = [
+  "General Sustainability",
+  "Food",
+  "Water",
+  "Energy",
+  "Transportation",
+  "Consumption & Waste",
+];
+
+/** @deprecated Use {@link ACTIONS} instead. Slated for removal in v0.8. */
+export const CATEGORIES = ACTIONS;
